Mix of changes from Partha
Cut down on some unnecessary mailbox consolidation
Improved display of the orange overlay when items in the inbox are expiring soon
Improved display of the tooltip
It is now possible to take the oldest item from a stack by holding down a modifier key. provided that the modifier key doesn't do something else first. (ctrl-click tends to preview armor pieces for instance)